Replaced getUserFromCookies with getCurrentUser
I decied to replace the getUserFromCookies function with a getCurrentUser function.
It makes more sense to provide the router handler the user record directly rather
than forcing it to query the database itself. Every route handler that needed the
user data in the datbase would have been forced to repeat the request.