Fixed missing JWT error on /me route
The JWT Cookie  was showing up as undefined when making requests to the /me route in
production. The cookie is accessible on every other production route. I added a line
to look at the header entries because the cookie was attached in my browser tools,
and for a reason I cannot discern, simply calling headers.entries() fixed the
problem. If I remove the invocation tnen I can no longer  access the cookie.
I don't know if this is a bug, or if I'm missing something about Next.js.

This problem was isolated to the "/api/me" route. The cookie for the JWT was showing
up as undefined despite working in my local enviornment and every other route in
production.