Built an auth guard higher order function
I was unhappy with using middleware to protect routes because I can't look at
at a specific route handler and determine if it is protected or not. I have to open
the middleware to find out, which is more confusing, so I decided to create a higher
order function that will tack an auth guard onto the handlers. I didn't want to copy
and paste the same auth check into every router handler, so this was the best option
for readability and maintainability. The auth guard checks if the current user is
authenticated and then feeds the current user into the route handler if all went
well.